+++ .gitmessage
@@ -1,4 +1,4 @@
-Feat: 사용자 삭제/비활성화 및 이를 위한 admin 모듈 개발
+Refactor: 스키마 정리
 # #로 시작하는 줄은 주석입니다. 주석은 지우지 마세요!!
 # <Type>(<scope>): <제목> (가능하면 50자 이내, 마침표 없이)
 # -----type 목록-----
