+++ src/backend/authentication/__init__.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timezone
-from pydantic import BaseModel
 from typing import Optional
 from itsdangerous import URLSafeTimedSerializer, BadSignature, SignatureExpired
 from argon2 import PasswordHasher
@@ -15,29 +14,6 @@ TOKEN_EXPIRE_SECONDS = 18000  # 5시간
 # 패스워드 해싱 및 인증 설정
 ph = PasswordHasher()
 serializer = URLSafeTimedSerializer(SECRET_KEY)
-
-class UserResponse(BaseModel):
-    username: str
-    disabled: bool = False
-    is_admin: bool = False
-    disabled_reason: Optional[str] = None
-    disabled_at: Optional[datetime] = None
-
-
-class UserCreate(BaseModel):
-    username: str
-    password: str
-
-
-class AccountAction(BaseModel):
-    """계정 활성화/비활성화를 위한 요청 모델"""
-    reason: Optional[str] = None
-
-
-class AdminUserAction(BaseModel):
-    """관리자의 사용자 관리 작업을 위한 요청 모델"""
-    username: str
-    reason: Optional[str] = None
 
 
 def hash_password(password: str) -> str:
